requirements.txt
ricecooker>=0.6.43
le_utils>=0.1.4
beautifulsoup4>=4.9.0
Pillow==2.2.2